Create test_utils.py
These are general utilities for testing Pokemon Terminal. The data and functions in this file can be used by other `test_*.py` scripts but they should not be imported into or copied into normal scripts in the project. This approach will help us to steer clear of assumption bias and avoid shared code between dev and test.